# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Steve Analiz.web API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  author: string;
  summary: string;
  content: string;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateBlogPostBody {
  title: string;
  slug: string;
  author: string;
  summary: string;
  content: string;
  tags: string[];
}

export type BlogStatsTagCounts = { [key: string]: number };

export interface BlogStats {
  totalPosts: number;
  tagCounts: BlogStatsTagCounts;
}

export interface ErrorResponse {
  error: string;
}

export type ListBlogPostsParams = {
  /**
   * Filter posts by tag
   */
  tag?: string;
};