# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Steve Analiz.web API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all blog posts ordered by newest first
 * @summary List all blog posts
 */
export const ListBlogPostsQueryParams = zod.object({
  tag: zod.coerce.string().optional().describe("Filter posts by tag"),
});

export const ListBlogPostsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  slug: zod.string(),
  author: zod.string(),
  summary: zod.string(),
  content: zod.string(),
  tags: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListBlogPostsResponse = zod.array(ListBlogPostsResponseItem);

/**
 * Adds a new blog post to the platform
 * @summary Create a new blog post
 */
export const CreateBlogPostBody = zod.object({
  title: zod.string(),
  slug: zod.string(),
  author: zod.string(),
  summary: zod.string(),
  content: zod.string(),
  tags: zod.array(zod.string()),
});

/**
 * @summary Get a single blog post by slug
 */
export const GetBlogPostParams = zod.object({
  slug: zod.coerce.string(),
});

export const GetBlogPostResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  slug: zod.string(),
  author: zod.string(),
  summary: zod.string(),
  content: zod.string(),
  tags: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * Returns counts by tag and total post count
 * @summary Get blog statistics
 */
export const GetBlogStatsResponse = zod.object({
  totalPosts: zod.number(),
  tagCounts: zod.record(zod.string(), zod.number()),
});